Remove .hfp from GitHublarge storage. Update/add some video/audio files
.hfp files were tiny so it didn't really make sense to store them using GitHub large storage.

Added mp3 of Battle Music 1 as the version in game is now only 10 minutes (as oppose to 60) and is a .ogg file.

Added Imperial March piano only as a .wav file, as that's what it was original when outputted from Hitfilm Expres (having been cropped). The version in game is a .flac file, smaller with the same sound quality, previously was an .mp3.

I now have the 10 minbutes exactly (previously 10 minutes 6 seconds...) of Sand Dunes.mp4. Version in game is a compressed .ogv file.