Export play store art/logos as webp images
This is to be used mainly on the website for install prompt screenshots. Given Chromium browser are the only browsers which install rich install prompts with images, and Chromium has supported webp since 2011 - way before rich install prompts or even PWAs, there should should be zero compatibility issues using it. To my knowledge the only format with better lossless compression than webp (I have compressed the pngs losslessly) is jpeg-xl, which is currently only supported by Safari/Webkit, so not an option at the currently.

Regardless, the webp outputs are around 44% (with variance from 37% to 72%) from the size of the pngs which is crazy, normally webp lossless is good but not that good. And as far I as can tell flicking between them side by side, they are identical.

Icons are not quite as clean because Safari on old macOS and old iOS devices (~0.8% combined marketshare, EOL pre-2020), old Firefox (-2018, 0.12%) and most kaiOS (~0.07%) don't support webp, but this is a pretty low number (1% overall) that's going to fall over time.