Modify Linux fuse to call executable love, final output Galaxy Collection
Previously both were Star Wars Galaxy Collection. The Star Wars probably wanted to go anyway, but I didn't want to change the executable to love, but it's the only way I can get it to run at the moment. I've also added the working AppStreamData to the Linux folder, which I've tried to attach to the AppImage but won't attach without the AppImage renaming and breaking it. And finally moved the Windows VersionInfo into the Window folder